# Conflicts:
#	browser/app/profile/firefox.js
#	browser/locales/en-US/browser/appmenu.ftl
#	caps/nsScriptSecurityManager.cpp
#	js/src/vm/JSFunction.cpp

+++ dom/media/webrtc/transport/nr_socket_prsock.h
@@ -310,6 +310,6 @@ int nr_praddr_to_transport_addr(const PRNetAddr* praddr,
                                 nr_transport_addr* addr, int protocol,
                                 int keep);
 int nr_transport_addr_get_addrstring_and_port(const nr_transport_addr* addr,
-                                              nsACString* host, int32_t* port);
+                                              nsACString* host, uint16_t* port);
 }  // namespace mozilla
 #endif
